+++ CPAC/longitudinal/wf/anat.py
@@ -43,6 +43,8 @@ from CPAC.utils.utils import check_prov_for_regtool
     name="mask_T1w_longitudinal_template",
     config=["longitudinal_template_generation"],
     switch=["run"],
+    option_key="using",
+    option_val="C-PAC legacy",
     inputs=["desc-brain_T1w"],
     outputs=["space-T1w_desc-brain_mask"],
 )
@@ -97,6 +99,8 @@ def select_session(
     name="mask_longitudinal_T1w_brain",
     config=["longitudinal_template_generation"],
     switch=["run"],
+    option_key="using",
+    option_val="C-PAC legacy",
     inputs=["space-longitudinal_desc-brain_T1w"],
     outputs=["space-longitudinal_desc-brain_mask"],
 )
@@ -122,6 +126,8 @@ def mask_longitudinal_T1w_brain(
     name="warp_longitudinal_T1w_to_template",
     config=["longitudinal_template_generation"],
     switch=["run"],
+    option_key="using",
+    option_val="C-PAC legacy",
     inputs=[
         (
             "space-longitudinal_desc-brain_T1w",
@@ -178,6 +184,8 @@ def warp_longitudinal_T1w_to_template(
     name="warp_longitudinal_seg_to_T1w",
     config=["longitudinal_template_generation"],
     switch=["run"],
+    option_key="using",
+    option_val="C-PAC legacy",
     inputs=[
         (
             "from-longitudinal_to-T1w_mode-image_desc-linear_xfm",
@@ -376,12 +384,14 @@ def anat_longitudinal_wf(
         template_node.inputs.set(
             avg_method=config.longitudinal_template_generation["average_method"],
             dof=config.longitudinal_template_generation["dof"],
-            interp=config.longitudinal_template_generation["interp"],
-            cost=config.longitudinal_template_generation["cost"],
+            interp=config.longitudinal_template_generation["legacy-specific"]["interp"],
+            cost=config.longitudinal_template_generation["legacy-specific"]["cost"],
             convergence_threshold=config.longitudinal_template_generation[
                 "convergence_threshold"
             ],
-            thread_pool=config.longitudinal_template_generation["thread_pool"],
+            thread_pool=config.longitudinal_template_generation["legacy-specific"][
+                "thread_pool"
+            ],
             unique_id_list=list(session_wfs.keys()),
         )